requirements3.7.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_pytorch_select=0.2=gpu_0
blas=1.0=mkl
boost=1.70.0=py37h9de70de_1
boost-cpp=1.70.0=ha2d47e9_1
bzip2=1.0.8=h516909a_2
ca-certificates=2020.6.20=hecda079_0
cairo=1.14.12=h8948797_3
cairomm=1.12.2=0
certifi=2020.6.20=py37hc8dfbb8_0
cffi=1.14.0=py37h2e261b9_0
cgal=4.14=py37he1b5a44_0
cgal-cpp=4.14.1=h261e690_0
cudatoolkit=10.0.130=0
cudnn=7.6.5=cuda10.0_0
cycler=0.10.0=py37_0
dbus=1.13.12=h746ee38_0
decorator=4.4.2=py_0
eigen=3.3.7=hc9558a2_1001
expat=2.2.6=he6710b0_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
glib=2.63.1=h5a9c865_0
gmp=6.2.0=he1b5a44_2
graph-tool=2.29=py37h3f2f939_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
icu=58.2=h9c2bf20_1
intel-openmp=2020.0=166
joblib=0.14.1=py_0
jpeg=9b=h024ee3a_2
kiwisolver=1.1.0=py37he6710b0_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libllvm8=8.0.1=hc9558a2_0
libpng=1.6.37=hbc83047_0
libsigcpp=2.10.0=hf484d3e_1001
libstdcxx-ng=9.1.0=hdf63c60_0
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.9=hea5a465_1
llvmlite=0.31.0=py37h5202443_1
matplotlib=3.1.3=py37_0
matplotlib-base=3.1.3=py37hef1b27d_0
mkl=2020.0=166
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.0.15=py37ha843d7b_0
mkl_random=1.1.0=py37hd6b4f25_0
mpfr=4.0.2=he80fd80_0
ncurses=6.2=he6710b0_0
networkx=2.4=py_0
ninja=1.9.0=py37hfd86e86_0
numba=0.48.0=py37hb3f55d8_0
numpy=1.18.1=py37h4f9e942_0
numpy-base=1.18.1=py37hde5b4d6_1
openssl=1.1.1g=h516909a_0
pandas=1.0.1=py37h0573a6f_0
pcre=8.43=he6710b0_0
pip=20.0.2=py37_1
pixman=0.38.0=h516909a_1003
pycairo=1.19.0=py37h2a1e443_0
pycparser=2.19=py37_0
pyparsing=2.4.6=py_0
pyqt=5.9.2=py37h05f1152_2
python=3.7.6=h0371630_2
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
pytorch=1.3.1=cuda100py37h53c1284_0
pytz=2019.3=py_0
qt=5.9.7=h5867ecd_1
readline=7.0=h7b6447c_5
scikit-learn=0.22.1=py37hd81dba3_0
scipy=1.4.1=py37h0b6359f_0
seaborn=0.10.0=py_0
setuptools=46.0.0=py37_0
sip=4.19.8=py37hf484d3e_0
six=1.14.0=py37_0
sparsehash=2.0.3=hf484d3e_1000
sqlite=3.31.1=h7b6447c_0
tbb=2020.1=hc9558a2_0
tk=8.6.8=hbc83047_0
tornado=6.0.4=py37h7b6447c_1
umap-learn=0.4.4=py37hc8dfbb8_0
wheel=0.34.2=py37_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=h7b6447c_3
